MaskPipeline/requirements.txt
numpy
opencv-python
PyYAML
imageio-ffmpeg
torch
torchvision
ultralytics
transformers
accelerate
